containerization/requirements.txt
PyYAML
Jinja2
boto3==1.35.95